Move essential matrix factors into SFM
