+++ gtsam/sfm/EssentialMatrixConstraint.cpp
@@ -16,7 +16,7 @@
  *  @date   Jan 5, 2014
  **/
 
-#include <gtsam/slam/EssentialMatrixConstraint.h>
+#include <gtsam/sfm/EssentialMatrixConstraint.h>
 //#include <gtsam/linear/GaussianFactor.h>
 //#include <gtsam/base/Testable.h>
